// src/types/scope-summary.ts
/*
 * Copyright 2020 EPAM Systems
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { ScopeCoverage } from "./scope-coverage";

export interface ScopeSummary {
  name: string;
  id: string;
  started: number;
  finished: number;
  enabled: boolean;
  active: boolean;
  coverage: ScopeCoverage;
  [key: string]: string | number | boolean | ScopeCoverage;
}